% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gg_animate.R
\name{animation_saver}
\alias{animation_saver}
\title{Retrieve a function for saving animations based on a string/function or a filename}
\usage{
animation_saver(saver, filename)
}
\arguments{
\item{saver}{A function or string describing an animation saver}

\item{filename}{File name to save to}
}
\description{
Retrieve a function for saving animations based on a string/function or a filename
}

