% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{auto_browse}
\alias{auto_browse}
\title{Auto browse to a filename}
\usage{
auto_browse(output)
}
\arguments{
\item{output}{Open a file}
}
\description{
This utility function is adapted from the animation package
\url{https://github.com/yihui/animation/blob/df12e57b3cb1a71a1935f5351e007d141af8ae2c/R/utils.R}
}

