% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ease-aes.R
\name{ease_aes}
\alias{ease_aes}
\title{Control easing of aesthetics}
\usage{
ease_aes(default, ...)
}
\arguments{
\item{default}{The default easing function to use}

\item{...}{Override easing for specific aesthetics}
}
\description{
Easing defines how a value change to another during tweening. Will it
progress linearly, or maybe start slowly and then build up momentum. In
\code{gganimate}, each aesthetic or computed variable can be tweened with
individual easing functions using the \code{ease_aes()} function. All easing
functions implemented in \code{tweenr} are available, see \link[tweenr:display_ease]{tweenr::display_ease}
for an overview.
}
