% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute-.R, R/transmute-enter.R,
%   R/transmute-exit.R
\name{enter_exit}
\alias{enter_exit}
\alias{enter}
\alias{exit}
\alias{enter_manual}
\alias{enter_appear}
\alias{enter_fade}
\alias{enter_grow}
\alias{exit_manual}
\alias{exit_disappear}
\alias{exit_fade}
\alias{exit_shrink}
\title{Define how entering and exiting data behaves}
\usage{
enter_manual(default, ...)

enter_appear(early = FALSE, ...)

enter_fade(..., alpha = 0)

enter_grow(..., size = 0, alpha = NA)

exit_manual(default, ...)

exit_disappear(early = FALSE, ...)

exit_fade(..., alpha = 0)

exit_shrink(..., size = 0, alpha = NA)
}
\arguments{
\item{default}{A default transformation to use}

\item{...}{Additional specific transformations either named by the geom
(e.g. \code{bar}, or by its position in the layer stack, e.g. \code{"2"})}

\item{early}{Should the data appear in the beginning of the transition or in
the end}

\item{alpha}{The start/end transparency. \code{alpha = NA} does not modify the
alpha values}

\item{size}{The proportional start/end size. \code{0} means complete shrinking
while \code{1} means no shrinking}
}
\description{
The purpose of \code{enter_*()} and \code{exit_*()} is to control what happens with
data that does not persist during a tween. In general the non-persistent data
is transformed to an \emph{invisible} version that can be tweened to, e.g. by
setting the opacity to 0 or be moving the element off-screen. It is possible
to define your own transformations, or rely on some of the build in
\emph{effects}.
}
\details{
For layers that are tweened based on the raw data, only the specified
aesthetics are available to modify, while all possible variables are
available for late-tweening layers

\strong{appear}/\strong{disappear} will simply make elements appear/disappear at either the
start or end of the transition.

\strong{fade} will simply set the alpha value to zero making the elements fade
in/out during the transition.

\strong{grow}/\strong{shrink} will set the elements to zero size making them gradually
grow into / shrink out of existence. Zero size depends on the type of layer,
e.g. polygons/paths will have all their points set to the mean, while points
will have size/stroke set to zero.
}
