% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{frame_vars}
\alias{frame_vars}
\title{Access metadata about the frames in an animation}
\usage{
frame_vars(animation = last_animation())
}
\arguments{
\item{animation}{The animation to extract metadata from. Defaults to the last
rendered animation}
}
\value{
A data.frame with a row per frame in the animation and metadata as
columns
}
\description{
This function extracts the metadata generated about each frame during
rendering. It corresponds to the information available to the labels for
string literal interpretation and is thus dependent on the transition and
view used for the animation.
}
