% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saving.R
\name{gg_animate_save}
\alias{gg_animate_save}
\title{Save a gganimate object to a file}
\usage{
gg_animate_save(g, filename = NULL, saver = NULL, ...)
}
\arguments{
\item{g}{A gg_animate object}

\item{filename}{File to write to}

\item{saver}{A string such as "mp4" or "gif" that specifies
a function from the animation package such as \code{saveVideo}
or \code{saveGIF} to use for saving. This can also be recognized from the
filename extension.}

\item{...}{Additional arguments passed on to the saving function,
such as \code{saveVideo} or \code{saveGIF} from the animation
package.}
}
\description{
Save a gganimate object to a file
}

