% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate-ggproto.R, R/shadow-null.R,
%   R/transition-.R, R/transition-null.R, R/transition-states.R, R/view-.R,
%   R/view-follow.R, R/view-step.R
\docType{data}
\name{gganimate-ggproto}
\alias{gganimate-ggproto}
\alias{ShadowNull}
\alias{Transition}
\alias{TransitionNull}
\alias{TransitionStates}
\alias{View}
\alias{ViewFollow}
\alias{ViewStep}
\title{Base ggproto classes for gganimate}
\description{
If you are creating a new transition, view, etc. in another package,
you'll need to extend from some of the ggproto classes defined in \code{gganimate}
}
\seealso{
ggproto
}
\keyword{datasets}
\keyword{internal}
