% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate-package.r
\docType{package}
\name{gganimate-package}
\alias{gganimate}
\alias{gganimate-package}
\title{gganimate: A Grammar of Animated Graphics}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The grammar of graphics as implemented in the 'ggplot2' package has
    been successful in providing a powerful API for creating static 
    visualisation. In order to extend the API for animated graphics this package
    provides a completely new set of grammar, fully compatible with 'ggplot2' 
    for specifying transitions and animations in a flexible and extensible way.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/thomasp85/gganimate}
  \item Report bugs at \url{http://github.com/thomasp85/gganimate/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

Authors:
\itemize{
  \item David Robinson \email{admiral.david@gmail.com}
}

}
