% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gganimate_tempfile}
\alias{gganimate_tempfile}
\title{Create a temporary file within the temporary directory}
\usage{
gganimate_tempfile(pattern = "file", fileext = "")
}
\arguments{
\item{pattern}{the initial part of the name}

\item{fileext}{file extension}
}
\description{
This is necessary because the animation package often copies
to the temporary directory, which leads to animation trying to
copy a file to itself.
}

