% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_type.R
\name{layer_type}
\alias{layer_type}
\title{Define what graphical type a layer is}
\usage{
layer_type(x)
}
\arguments{
\item{x}{a Layer, Geom, or Stat}
}
\description{
Define what graphical type a layer is
}
\keyword{internal}
