% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_modifiers.R
\name{outer_align}
\alias{outer_align}
\title{Align labels with outer edge of the image in a gganimate animation}
\usage{
outer_align(plot)
}
\arguments{
\item{plot}{a plot formatted as a gtable}
}
\value{
a plot formatted as a gtable
}
\description{
The function takes a gtable object and returns a gtable object,
with the title, subtitle, and caption aligned to the outer boundary of the image.
This function is not intended to be called directly - it is intended to be
supplied to the \code{frame_modifier} argument of \code{animate()}.
}
\examples{

plot <- ggplot(mtcars, aes(x = wt, y = mpg)) +
   geom_point() +
   labs(title = "Title",
        subtitle = "Subtitle",
        caption = "Caption")

anim_plot <- plot +
   transition_states(gear, transition_length = 2, state_length = 1) +
   enter_fade() +
   exit_fade()

\dontrun{animate(anim_plot, frame_modifier = outer_align)}

}
