% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_ggplot_build}
\alias{plot_ggplot_build}
\title{Plot a built ggplot object}
\usage{
plot_ggplot_build(b, newpage = is.null(vp), vp = NULL)
}
\arguments{
\item{b}{A list resulting from \code{\link{ggplot_build}}}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}
}
\description{
We needed a customized version of ggplot2's \code{print.ggplot2},
because we need to build plots from the intermediate results of
\code{\link{ggplot_build}} rather than from a \code{gg} object.
}

