% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate.R
\name{print.gganimate}
\alias{print.gganimate}
\title{Print a gganimate object, allowing browsing in RStudio}
\usage{
\method{print}{gganimate}(x, format = "gif", ...)
}
\arguments{
\item{x}{gganimate object}

\item{format}{What format to display in, such as "gif" (default),
"mp4", or "avi".}

\item{...}{Extra arguments for the <img> or <video> tag, such
as width or height

This saves the plot to a file using \code{\link{gganimate_save}}
(and then loads the contents of that file into memory) if it has
not already been saved.}
}
\description{
Print a gganimate object as browsable HTML, which allows visualization
directly in RStudio. If we are in knitr, directly print each of the
images instead (you should use the \code{fig.show = "animate"} option
in the chunk).
}

