% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim_save.R
\name{save_animation}
\alias{save_animation}
\title{Infrastructure for animation saving}
\usage{
save_animation(animation, file)
}
\arguments{
\item{animation}{The animation object to save}

\item{file}{The file path to save it to}
}
\description{
Any returned animation object that wish to support \code{\link[=anim_save]{anim_save()}} need to
implement this method
}
\keyword{internal}
