% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-null.R
\name{shadow_null}
\alias{shadow_null}
\title{A non-existent shadow}
\usage{
shadow_null()
}
\description{
This is the default shadow that simply doesn't show anything other than the
data for the current frame.
}
\seealso{
Other shadows: 
\code{\link{shadow_mark}()},
\code{\link{shadow_trail}()},
\code{\link{shadow_wake}()}
}
\concept{shadows}
