% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-wake.R
\name{shadow_wake}
\alias{shadow_wake}
\title{Show preceeding frames with gradual falloff}
\usage{
shadow_wake(wake_length, size = TRUE, alpha = TRUE, falloff = "cubic-in",
  wrap = TRUE, exclude_layer = NULL)
}
\arguments{
\item{wake_length}{A number between 0 and 1 giving the length of the wake,
in relation to the total number of frames.}

\item{size}{Boolean indicating whether the size of the geom should shrink}

\item{alpha}{Boolean indicating whether geoms should get more translucent}

\item{falloff}{An easing function that control how size and/or alpha should
change.}

\item{wrap}{Should the shadow wrap around, so that the first frame will get
shadows from the end of the animation.}

\item{exclude_layer}{Indexes of layers that should be excluded.}
}
\description{
This shadow is meant to draw a small wake after data by showing the latest
frames up to the current. You can choose to gradually diminish the size
and/or opacity of the shadow. The length of the wake is not given in absolute
frames as that would make the animation susceptible to changes in the
framerate. Instead it is given as a proportion of the total length of the
animation.
}
\seealso{
Other shadows: \code{\link{shadow_null}}
}
\concept{shadows}
