% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-components.R
\name{transition_components}
\alias{transition_components}
\title{Transition individual components through their own lifecycle}
\usage{
transition_components(time, range = NULL, enter_length = NULL,
  exit_length = NULL, id)
}
\arguments{
\item{time}{The unquoted name of the column holding the time for each state
of the components}

\item{range}{The range the animation should span. Defaults to the range of
time plus enter and exit length}

\item{enter_length, exit_length}{How long time should be spend on enter and
exit transitions. Defaults to 0}

\item{id}{\strong{Deprecated}}
}
\description{
This transition allows individual visual components to define their own
"life-cycle". This means that the final animation will not have any common
"state" and "transition" phase as any component can be moving or static at
any point in time.
}
\section{Label variables}{

\code{transition_components} makes the following variables available for string
literal interpretation:
\itemize{
\item \strong{frame_time} gives the time that the current frame corresponds to
}
}

\section{Object permanence}{

\code{transition_components} uses the group aesthetic of each layer to identify
which rows in the input data correspond to the same graphic element and will
therefore define stages in time that the element will animate through. The
group aesthetic, if not set, will be calculated from the interaction of all
discrete aesthetics in the layer (excluding \code{label}), so it is often better
to set it explicetly when animating, to make sure your data is interpreted in
the right way. If the group aesthetic is not set, and no discrete aesthetics
exists then all rows will have the same group.
}

\seealso{
Other transitions: \code{\link{transition_events}},
  \code{\link{transition_filter}},
  \code{\link{transition_layers}},
  \code{\link{transition_manual}},
  \code{\link{transition_null}},
  \code{\link{transition_reveal}},
  \code{\link{transition_states}},
  \code{\link{transition_time}}
}
\concept{transitions}
