% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-events.R
\name{transition_events}
\alias{transition_events}
\title{Transition individual events in and out}
\usage{
transition_events(start, end = NULL, range = NULL,
  enter_length = NULL, exit_length = NULL)
}
\arguments{
\item{start, end}{The unquoted expression giving the start and end time of
each event. If \code{end}is \code{NULL} the event will be treated as having no duration.}

\item{range}{The range the animation should span. Defaults to the range of
the events from they enter to they have exited.}

\item{enter_length, exit_length}{The unquoted expression giving the length to
be used for enter and exit for each event.}
}
\description{
This transition treats each visual element as an event in time and allows you
to control the duration and enter/exit length individually for each event.
}
\section{Label variables}{

\code{transition_components} makes the following variables available for string
literal interpretation:
\itemize{
\item \strong{frame_time} gives the time that the current frame corresponds to
}
}

\seealso{
Other transitions: \code{\link{transition_components}},
  \code{\link{transition_filter}},
  \code{\link{transition_layers}},
  \code{\link{transition_manual}},
  \code{\link{transition_null}},
  \code{\link{transition_reveal}},
  \code{\link{transition_states}},
  \code{\link{transition_time}}
}
\concept{transitions}
