% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-filter.R
\name{transition_filter}
\alias{transition_filter}
\title{Transition between different filters}
\usage{
transition_filter(transition_length, filter_length, ..., wrap = TRUE,
  keep = FALSE)
}
\arguments{
\item{transition_length}{The relative length of the transition. Will be
recycled to match the number of states in the data}

\item{filter_length}{The relative length of the pause at the states. Will be
recycled to match the number of states in the data}

\item{...}{A number of expressions to be evaluated in the context of the layer
data, returning a logical vector. If the expressions are named, the name will
be available as a frame variable.}

\item{wrap}{Should the animation \emph{wrap-around}? If \code{TRUE} the last filter will
be transitioned into the first.}

\item{keep}{Should rows that evaluates to \code{FALSE} be kept in the data as it
looks after exit has been applied}
}
\description{
This transition allows you to transition between a range of filtering
conditions. The conditions are expressed as logical statements and rows in
the data will be retained if the statement evaluates to \code{TRUE}. It is
possible to keep filtered data on display by setting \code{keep = TRUE} which will
let data be retained as the result of the exit function. Note that if data is
kept the enter function will have no effect.
}
\section{Label variables}{

\code{transition_filter} makes the following variables available for string
literal interpretation:
\itemize{
\item \strong{transitioning} is a boolean indicating whether the frame is part of the
transitioning phase
\item \strong{previous_filter} The name of the last filter the animation was at
\item \strong{closest_filter} The name of the filter closest to this frame
\item \strong{next_filter} The name of the next filter the animation will be part of
\item \strong{previous_expression} The expression of the last filter the animation was at
\item \strong{closest_expression} The expression of the filter closest to this frame
\item \strong{next_expression} The expression of the next filter the animation will be part of
}
}

\seealso{
Other transitions: \code{\link{transition_components}},
  \code{\link{transition_events}},
  \code{\link{transition_layers}},
  \code{\link{transition_manual}},
  \code{\link{transition_null}},
  \code{\link{transition_reveal}},
  \code{\link{transition_states}},
  \code{\link{transition_time}}
}
\concept{transitions}
