% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-layer.R
\name{transition_layers}
\alias{transition_layers}
\title{Build up a plot, layer by layer}
\usage{
transition_layers(layer_length, transition_length, keep_layers = TRUE,
  from_blank = TRUE)
}
\arguments{
\item{layer_length}{The proportional time to pause at each layer before a new
one enters}

\item{transition_length}{The proportional time to usse for the entrance of a
new layer}

\item{keep_layers}{Should layers be kept on screen after they have appeeared
or transition out when a new layer enters}

\item{from_blank}{Should the first layer transition in or be present on the
onset of the animation}
}
\description{
This transition gradually adds layers to the plot in the order they have been
defined. By default prior layers are kept for the remainder of the animation,
but they can also be set to be removed as the next layer enters.
}
\seealso{
Other transitions: \code{\link{transition_null}},
  \code{\link{transition_states}}
}
\concept{transitions}
