% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-manual.R
\name{transition_manual}
\alias{transition_manual}
\title{Create an animation by specifying the frame membership directly}
\usage{
transition_manual(frames, ..., cumulative = FALSE)
}
\arguments{
\item{frames}{The unquoted name of the column holding the frame membership.}

\item{...}{Additional variables}

\item{cumulative}{Keep data from previous frames as part of the current frame
data}
}
\description{
This transition allows you to map a variable in your data to a specific frame
in the animation. No tweening of data will be made and the number of frames
in the animation will be decided by the number of levels in the frame
variable.
}
\section{Label variables}{

\code{transition_states} makes the following variables available for string
literal interpretation:
\itemize{
\item \strong{previous_frame} The name of the last frame the animation was at
\item \strong{current_frame} The name of the current frame
\item \strong{next_frame} The name of the next frame in the animation
}
}

\seealso{
Other transitions: \code{\link{transition_components}},
  \code{\link{transition_events}},
  \code{\link{transition_filter}},
  \code{\link{transition_layers}},
  \code{\link{transition_null}},
  \code{\link{transition_reveal}},
  \code{\link{transition_states}},
  \code{\link{transition_time}}
}
\concept{transitions}
