% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-null.R
\name{transition_null}
\alias{transition_null}
\title{Keep all data constant across the animation}
\usage{
transition_null()
}
\description{
Keep all data constant across the animation
}
\seealso{
Other transitions: \code{\link{transition_states}}
}
\concept{transitions}
