% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-time.R
\name{transition_time}
\alias{transition_time}
\title{Transition through distinct states in time}
\usage{
transition_time(time, range = NULL)
}
\arguments{
\item{time}{An unquoted expression giving the time, and thus state
membership, of each observation.}

\item{range}{The time range to animate. If \code{NULL} it will be set to the range
of \code{time}}
}
\description{
This is a variant of \code{\link[=transition_states]{transition_states()}} that is intended for data where
the states are representing specific point in time. The transition length
between the states will be set to correspond to the actual time difference
between them.
}
\section{Label variables}{

\code{transition_time} makes the following variables available for string
literal interpretation:
\itemize{
\item \strong{frame_time} gives the time that the current frame corresponds to
}
}

\examples{

p <- ggplot(airquality, aes(Day, Temp)) +
  geom_line(color = 'red', size = 1) +
  transition_time(Month)

# animate(p)

}
\seealso{
Other transitions: \code{\link{transition_components}},
  \code{\link{transition_events}},
  \code{\link{transition_filter}},
  \code{\link{transition_layers}},
  \code{\link{transition_manual}},
  \code{\link{transition_null}},
  \code{\link{transition_reveal}},
  \code{\link{transition_states}}
}
\concept{transitions}
