% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween_before_stat.R
\name{tween_before_stat}
\alias{tween_before_stat}
\title{Define if layers should be tweened before stats are calculated}
\usage{
tween_before_stat(x)
}
\arguments{
\item{x}{a Layer, Geom, or Stat}
}
\description{
Define if layers should be tweened before stats are calculated
}
\keyword{internal}
