% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-static.R
\name{view_static}
\alias{view_static}
\title{Keep a fixed view that include all of the data}
\usage{
view_static()
}
\description{
This view keeps positional scales fixed across all frames
}
\seealso{
Other views: 
\code{\link{view_follow}()},
\code{\link{view_step}()},
\code{\link{view_zoom}()}
}
\concept{views}
