% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{animation_saver}
\alias{animation_saver}
\title{Retrieve a function for saving animations based on a string/function or a filename}
\usage{
animation_saver(saver, filename, mime_type = NULL)
}
\arguments{
\item{saver}{A function or string describing an animation saver}

\item{filename}{File name to save to}

\item{mime_type}{If saver is a custom function instead of a string
specification, can specify a mime_type to save it as. Without this,
files can be saved but not browsed in RStudio.}
}
\description{
Retrieve a function for saving animations based on a string/function or a filename
}

