% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_animate.R
\name{gg_animate}
\alias{gg_animate}
\title{Show an animation of a ggplot2 object}
\usage{
gg_animate(p = last_plot(), filename = NULL, saver = NULL, pause = NULL,
  title_frame = TRUE, ...)
}
\arguments{
\item{p}{A ggplot2 object. If no plot is provided, use the last plot by default.}

\item{filename}{Output file. If not given, simply prints to the screen (typical for animated
knitr chunks)}

\item{saver}{Can specify a function (or a string such as "mp4" or "html" that specifies
a function to use for saving}

\item{pause}{Amount of time to pause between displaying each plot. Only used when
displaying to the screen, not saving to a file (and not useful when creating an
animation in a knitr chunk). When saving to a file, use
\code{ani.options(interval = ...)}}

\item{title_frame}{Whether to title each image with the current \code{frame} value.}

\item{...}{If saving to a file, extra arguments to pass along to the animation
saving function (to \code{saveVideo}/\code{saveGIF}/etc)}
}
\description{
Show an animation of a ggplot2 object that contains a \code{frame} aesthetic. This
\code{frame} aesthetic will determine which frame the animation is shown in. For
example, you could add the aesthetic \code{frame = time} to a dataset including
a \code{time} variable.
}

