% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate-ggproto.R, R/shadow-.R,
%   R/shadow-null.R, R/shadow-stamp.R, R/shadow-wake.R, R/transition-.R,
%   R/transition-manual.R, R/transition-components.R, R/transition-events.R,
%   R/transition-layers.R, R/transition-null.R, R/transition-states.R,
%   R/transition-time.R, R/view-.R, R/view-follow.R, R/view-step.R
\docType{data}
\name{gganimate-ggproto}
\alias{gganimate-ggproto}
\alias{Shadow}
\alias{ShadowNull}
\alias{ShadowMark}
\alias{ShadowWake}
\alias{Transition}
\alias{TransitionManual}
\alias{TransitionComponents}
\alias{TransitionEvents}
\alias{TransitionLayers}
\alias{TransitionNull}
\alias{TransitionStates}
\alias{TransitionTime}
\alias{View}
\alias{ViewFollow}
\alias{ViewStep}
\title{Base ggproto classes for gganimate}
\description{
If you are creating a new transition, view, etc. in another package,
you'll need to extend from some of the ggproto classes defined in \code{gganimate}
}
\seealso{
ggproto
}
\keyword{datasets}
\keyword{internal}
