% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate-ggproto.R, R/shadow-.R,
%   R/shadow-mark.R, R/shadow-null.R, R/shadow-trail.R, R/shadow-wake.R,
%   R/transition-.R, R/transition-components.R, R/transition-events.R,
%   R/transition-filter.R, R/transition-manual.R, R/transition-layers.R,
%   R/transition-null.R, R/transition-states.R, R/transition-time.R,
%   R/transition_reveal.R, R/view-.R, R/view-follow.R, R/view-step.R,
%   R/view-step-manual.R, R/view-zoom.R, R/view-zoom-manual.R
\docType{data}
\name{gganimate-ggproto}
\alias{gganimate-ggproto}
\alias{Shadow}
\alias{ShadowMark}
\alias{ShadowNull}
\alias{ShadowTrail}
\alias{ShadowWake}
\alias{Transition}
\alias{TransitionComponents}
\alias{TransitionEvents}
\alias{TransitionFilter}
\alias{TransitionManual}
\alias{TransitionLayers}
\alias{TransitionNull}
\alias{TransitionStates}
\alias{TransitionTime}
\alias{TransitionReveal}
\alias{View}
\alias{ViewFollow}
\alias{ViewStep}
\alias{ViewStepManual}
\alias{ViewZoom}
\alias{ViewZoomManual}
\title{Base ggproto classes for gganimate}
\description{
If you are creating a new transition, view, etc. in another package,
you'll need to extend from some of the ggproto classes defined in \code{gganimate}
}
\seealso{
ggproto
}
\keyword{datasets}
\keyword{internal}
