% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_column.R, R/layer_type.R,
%   R/match_shapes.R
\name{group_column}
\alias{group_column}
\alias{layer_type}
\alias{match_shapes}
\title{Define what graphical type a layer is}
\usage{
group_column(x)

layer_type(x)

match_shapes(x)
}
\arguments{
\item{x}{a Layer, Geom, or Stat}
}
\description{
Define what graphical type a layer is
}
\keyword{internal}
