% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderers.R
\name{renderers}
\alias{renderers}
\alias{magick_renderer}
\alias{file_renderer}
\title{Renderers provided by gganimate}
\usage{
magick_renderer()

file_renderer(dir = "~", prefix = "gganim_plot")
}
\arguments{
\item{dir}{The directory to copy the frames to}

\item{prefix}{The filename prefix to use for the image files}
}
\value{
The provided renderers are factory functions that returns a new function
that take \code{frames} and \code{fps} as arguments, the former being a character
vector with file paths to the images holding the separate frames, in the
order they should appear, and the latter being the framerate to use for the
animation in frames-per-second.
}
\description{
The purpose of the renderer function is to take a list of image files and
assemble them into an animation. \code{gganimate} provide a range of renderers
but it is also possible to provide your own, if the supplied ones are lacking
in any way. A renderer is given as argument to \code{\link[=animate]{animate()}}/print() and
recieves the paths to the individual frames once they have been created.
}
\details{
The return value of a rendering is dependent on the specific
renderer. \code{magick_renderer()} returns a \code{magick-image} invisibly and displays
the animation in the viewer. \code{file_renderer()} returns the path to the image
files invisibly and does not display anything.

It is possible to provide your own renderer function providing that it
matches the required signature (\code{frames} and \code{fps} argument). The return
value of your provided function will be the return value ultimately given by
\code{\link[=animate]{animate()}}
}
