% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-components.R
\name{transition_components}
\alias{transition_components}
\title{Transition individual components through their own lifecycle}
\usage{
transition_components(id, time, range = NULL, enter_length = NULL,
  exit_length = NULL)
}
\arguments{
\item{id}{The unquoted name of the column holding the id that links
components across the data}

\item{time}{The unquoted name of the column holding the time for each state
of the components}

\item{range}{The range the animation should span. Defaults to the range of
time plus enter and exit length}

\item{enter_length, exit_length}{How long time should be spend on enter and
exit transitions. Defaults to 0}
}
\description{
This transition allows individual visual components to define their own
"life-cycle". This means that the final animation will not have any commen
"state" and "transition" phase as any component can be moving or static at
any point in time.
}
\section{Label variables}{

\code{transition_components} makes the following variables available for string
literal interpretation:
\itemize{
\item \strong{frame_time} gives the time that the current frame corresponds to
}
}

\seealso{
Other transitions: \code{\link{transition_evetns}},
  \code{\link{transition_layers}},
  \code{\link{transition_manual}},
  \code{\link{transition_null}},
  \code{\link{transition_states}}
}
\concept{transitions}
