% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-layers.R
\name{transition_layers}
\alias{transition_layers}
\title{Build up a plot, layer by layer}
\usage{
transition_layers(layer_length, transition_length, keep_layers = TRUE,
  from_blank = TRUE, layer_names = NULL)
}
\arguments{
\item{layer_length}{The proportional time to pause at each layer before a new
one enters}

\item{transition_length}{The proportional time to usse for the entrance of a
new layer}

\item{keep_layers}{Should layers be kept on screen after they have appeeared
or transition out when a new layer enters}

\item{from_blank}{Should the first layer transition in or be present on the
onset of the animation}

\item{layer_names}{A character vector of names for each layers, to be used
when interpreting label literals}
}
\description{
This transition gradually adds layers to the plot in the order they have been
defined. By default prior layers are kept for the remainder of the animation,
but they can also be set to be removed as the next layer enters.
}
\section{Label variables}{

\code{transition_layers} makes the following variables available for string
literal interpretation:
\itemize{
\item \strong{transitioning} is a booloean indicating whether the frame is part of the
transitioning phase
\item \strong{previous_layer} The name of the last layer the animation was showing
\item \strong{closest_layer} The name of the layer the animation is closest to showing
\item \strong{next_layer} The name of the next layer the animation will show
\item \strong{nlayers} The total number of layers
}
}

\seealso{
Other transitions: \code{\link{transition_null}},
  \code{\link{transition_states}}
}
\concept{transitions}
