% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-null.R
\name{transition_null}
\alias{transition_null}
\title{Keep all data constant across the animation}
\usage{
transition_null()
}
\description{
Keep all data constant across the animation
}
\seealso{
Other transitions: 
\code{\link{transition_components}()},
\code{\link{transition_events}()},
\code{\link{transition_filter}()},
\code{\link{transition_layers}()},
\code{\link{transition_manual}()},
\code{\link{transition_reveal}()},
\code{\link{transition_states}()},
\code{\link{transition_time}()}
}
\concept{transitions}
