% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-follow.R
\name{view_follow}
\alias{view_follow}
\title{Let the view follow the data in each frame}
\usage{
view_follow()
}
\description{
This view will set the panels to include the data present in the frame.
}
\seealso{
Other views: \code{\link{view_static}},
  \code{\link{view_step}}, \code{\link{view_zoom}}
}
\concept{views}
